refactor device tree module into seperate submodules
